requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
gspread==5.7.2
oauth2client==4.1.3
requests==2.28.1
transitions==0.9.0
line-bot-sdk==2.3.0
graphviz==0.20.1
firebase-admin==6.0.1